v0.4.1 - Allow for multiple simultaneous instances
In 0.4.0 there was a bug if you tried to run multiple Resin instances
simultaneously. The implementation of the state-keeping Agent was a bit too
naive and thought it would be the only one, ever. Resin now makes sure to
actually store the Pid of its corresponding agent in the `options` keyword list
that `Resin.init/1` returns, so that it can keep track of what agent to
`PerformanceForecast.pop/1` in `Resin.call/2`.